moved `perspective-editor` and `widget-editor` in seperate folder
