+++ assets/js/src/core/bootstrap.ts
@@ -26,6 +26,7 @@ import '@Pimcore/modules/execution-engine'
 import '@Pimcore/modules/element/dynamic-types'
 import '@Pimcore/modules/app/theme/dynamic-types'
 import '@Pimcore/modules/wysiwyg'
+import '@Pimcore/modules/perspective-editor'
 import '@Pimcore/modules/widget-editor'
 import 'flexlayout-react/style/light.css'
 import '../../../css/globals.css'
