+++ assets/js/src/core/modules/perspective-editor/components/perspective-detail/perspective-detail-container.tsx
