+++ assets/js/src/core/modules/perspective-editor/components/perspective-form/components/allowed-menu-entries-panel/allowed-menu-entries-panel.styles.ts
