+++ assets/js/src/core/modules/perspective-editor/components/perspective-form/components/widget-configurator/components/stacked-widget-list/components/right-toolbar/right-toolbar.tsx
