+++ assets/js/src/core/modules/perspective-editor/components/perspective-form/components/widget-configurator/components/stacked-widget-list/stacked-widget-list.tsx
