+++ assets/js/src/core/modules/perspective-editor/components/perspective-form/components/widget-configurator/components/widget-configuration-card-item/widget-configuration-card-item.tsx
