+++ assets/js/src/core/modules/perspective-editor/components/perspective-form/components/widget-configurator/components/widget-configuraton-card/widget-configuration-card.tsx
