+++ assets/js/src/core/modules/perspective-editor/components/perspective-form/components/widget-configurator/context/hooks/use-widget-configurator-context.tsx
