+++ assets/js/src/core/modules/perspective-editor/components/perspective-form/components/widget-configurator/context/widget-configurator-provider.tsx
