+++ assets/js/src/core/modules/perspective-editor/components/perspective-form/components/widget-configurator/widget-configurator.tsx
