+++ assets/js/src/core/modules/perspective-editor/hooks/use-perspective-editor.tsx
