+++ assets/js/src/core/modules/perspective-editor/perspective-editor-container-inner.tsx
