+++ assets/js/src/core/modules/perspective-editor/registry/widget-registry.ts
