+++ assets/js/src/core/modules/widget-editor/components/tree/tree-container.tsx
