+++ assets/js/src/core/modules/widget-editor/components/widget-detail/widget-detail-container.tsx
