+++ assets/js/src/core/modules/widget-editor/components/widget-type-form/components/allowed-context-menu-panel/allowed-context-menu.styles.ts
