+++ assets/js/src/core/modules/widget-editor/components/widget-type-form/components/allowed-context-menu-panel/data-object-context-menu-options-panel.tsx
