+++ assets/js/src/core/modules/widget-editor/components/widget-type-form/components/allowed-context-menu-panel/document-context-menu-options-panel.tsx
