+++ assets/js/src/core/modules/widget-editor/components/widget-type-form/components/allowed-objects-panel/allowed-objects-panel.tsx
