+++ assets/js/src/core/modules/widget-editor/components/widget-type-form/components/filter-panel/filter-panel.tsx
