+++ assets/js/src/core/modules/widget-editor/components/widget-type-form/components/general-tab/general-tab.tsx
