+++ assets/js/src/core/modules/widget-editor/components/widget-type-form/element-tree-widget-type-form.tsx
