+++ assets/js/src/core/modules/widget-editor/components/widget-type-form/hooks/use-widget-type-form.ts
