+++ assets/js/src/core/modules/widget-editor/components/widget-type-form/widget-form.tsx
