+++ assets/js/src/core/modules/widget-editor/dynmic-types/definitions/dynamic-type-widget-type-abstract.tsx
