+++ assets/js/src/core/modules/widget-editor/dynmic-types/registry/dynamic-type-widget-type-registry.tsx
