+++ assets/js/src/core/modules/widget-editor/hooks/use-widget-editor.tsx
