+++ assets/js/src/core/modules/widget-editor/index.ts
@@ -8,31 +8,24 @@
  *  @license    Pimcore Open Core License (POCL)
  */
 
-import { serviceIds } from '@Pimcore/app/config/services/service-ids'
 import { container } from '@Pimcore/app/depency-injection'
 import { moduleSystem } from '@Pimcore/app/module-system/module-system'
-import { type MainNavRegistry } from '../app/base-layout/main-nav/services/main-nav-registry'
-import { type WidgetRegistry } from '../widget-manager/services/widget-registry'
-import { PerspectiveEditorContainer } from './perspective-editor/perspective-editor-container'
-import { WidgetEditorContainer } from './custom-view-editor/widget-editor-container'
-import '@Pimcore/modules/widget-editor/custom-view-editor'
+import { serviceIds } from '@sdk/app'
+import { type DynamicTypeWidgetTypeElementTree } from './dynmic-types/definitions/dynamic-type-widget-type-element-tree'
+import { type DynamicTypeWidgetTypeRegistry } from './dynmic-types/registry/dynamic-type-widget-type-registry'
+import { WidgetRegistry } from '../widget-manager/services/widget-registry'
+import { WidgetEditorContainer } from './widget-editor-container'
+import { MainNavRegistry } from '../app/base-layout/main-nav/services/main-nav-registry'
 
 moduleSystem.registerModule({
   onInit: () => {
     const widgetRegistryService = container.get<WidgetRegistry>(serviceIds.widgetManager)
-
-    widgetRegistryService.registerWidget({
-      name: 'perspective-editor',
-      component: PerspectiveEditorContainer
-    })
-
     widgetRegistryService.registerWidget({
       name: 'widget-editor',
       component: WidgetEditorContainer
     })
 
     const mainNavRegistryService = container.get<MainNavRegistry>(serviceIds.mainNavRegistry)
-
     mainNavRegistryService.registerMainNavItem({
       path: 'System/Widget-Editor',
       label: 'navigation.widget-editor.widget-editor',
@@ -54,26 +47,9 @@ moduleSystem.registerModule({
       }
     })
 
-    mainNavRegistryService.registerMainNavItem({
-      path: 'System/Perspective-Editor',
-      label: 'navigation.widget-editor.perspective-editor',
-      order: 400,
-      dividerBottom: true,
-      className: 'item-style-modifier',
-      // permission: UserPermission.FOO,
-      // perspectivePermission: NavPermission.BAR,
-      widgetConfig: {
-        name: 'perspectiveEditor',
-        id: 'perspective-editor',
-        component: 'perspective-editor',
-        config: {
-          translationKey: 'widget.widget-editor.perspective-editor',
-          icon: {
-            type: 'name',
-            value: 'book-open-01'
-          }
-        }
-      }
-    })
+    const widgetRegistry = container.get<DynamicTypeWidgetTypeRegistry>(serviceIds['DynamicTypes/WidgetEditor/WidgetTypeRegistry'])
+    widgetRegistry.registerDynamicType(
+      container.get<DynamicTypeWidgetTypeElementTree>(serviceIds['DynamicTypes/WidgetEditor/ElementTree'])
+    )
   }
 })
