+++ assets/js/src/core/modules/widget-editor/widget-editor-container-inner.tsx
